1. Updated StartIndex from uint32 to uint64 2. Added new test cases.

+++ prover/server/prover/test_data_helpers.go
@@ -362,7 +362,7 @@ func BuildTestBatchAppendWithProofsTree(treeDepth int, batchSize int, previousTr
 		Height:              uint32(treeDepth),
 		BatchSize:           uint32(batchSize),
 		Tree:                &tree,
-		StartIndex:          uint32(startIndex),
+		StartIndex:          uint64(startIndex),
 	}
 }
